Add static_cast int->size_t for array index
Required by the CL:
Add "'cobalt_code': 1" to dom.gyp, fix Windows build
https://lbshell-internal-review.git.corp.google.com/#/c/49440/

+++ media/base/ranges.h
@@ -119,12 +119,12 @@ size_t Ranges<T>::size() const {
 
 template<class T>
 T Ranges<T>::start(int i) const {
-  return ranges_[i].first;
+  return ranges_[static_cast<size_t>(i)].first;
 }
 
 template<class T>
 T Ranges<T>::end(int i) const {
-  return ranges_[i].second;
+  return ranges_[static_cast<size_t>(i)].second;
 }
 
 template<class T>
